+++ pywallhaven/__init__.py
@@ -1,1 +1,2 @@
+# noinspection PyUnresolvedReferences
 from pywallhaven.wallhaven import *